Move all data to ViewTypeOfPerson
Move all contents from L5 to a new view called TypeOfPerson
